4.21 done

+++ controllers/blogs.js
@@ -31,6 +31,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(400).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  console.log(user)
+  console.log(blog)
+  
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(400).json({ error: 'you don\'t have permission to remove this blog' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end()
 })
